feat: Refactor and implement BN254 Groth16 Arkworks serializer; update related tests and build scripts
