+++ sui/.gitignore
@@ -20,3 +20,6 @@ bun.lockb
 
 # Environment-specific package overrides
 suigen-config.local.json
+
+# WASM build artifacts
+circuits/bn254-groth16-arkworks-serializer/
