+++ sui/circuits/.gitignore
@@ -7,8 +7,8 @@ keys/*.zkey
 verifier.sol
 
 # Rust build artifacts
-convert-vk/target/
-convert-vk/Cargo.lock
+bn254-groth16-arkworks-serializer/target/
+bn254-groth16-arkworks-serializer/Cargo.lock
 
 # Generated intermediate files
 keys/*_arkworks.json
