+++ sui/scripts/build-wasm.sh
@@ -1,9 +1,9 @@
 #!/usr/bin/env bash
 set -euo pipefail
 
-echo "🔨 Building WASM module for Arkworks converter..."
+echo "🔨 Building WASM module for BN254 Groth16 Arkworks Serializer..."
 
-cd circuits/convert-vk
+cd circuits/bn254-groth16-arkworks-serializer
 
 # wasm-packインストール確認
 if ! command -v wasm-pack &> /dev/null; then
@@ -15,9 +15,9 @@ fi
 echo "🏗️ Building WASM..."
 wasm-pack build \
     --target web \
-    --out-dir ../../public/wasm/arkworks-converter \
-    --out-name arkworks-converter
+    --out-dir ../../public/wasm/bn254-groth16-arkworks-serializer \
+    --out-name bn254_groth16_arkworks_serializer
 
 echo "✅ WASM module built successfully"
-echo "📍 Output: public/wasm/arkworks-converter/"
-ls -lh ../../public/wasm/arkworks-converter/
+echo "📍 Output: public/wasm/bn254-groth16-arkworks-serializer/"
+ls -lh ../../public/wasm/bn254-groth16-arkworks-serializer/
